ext_man: Add element type enumeration in rimage
Such a split will allow to add dictionary elements with list
of elements placed in particular elf section - because of possibility
to calculate sizeof of this section.
Moreover with such an approach, it is possible to define list of
mandatory extended manifest elements, which will be checked after
each build for each supported platform.

Signed-off-by: Karol Trzcinski <karolx.trzcinski@linux.intel.com>