rimage: correct module type
Currently some module types are incorrect and it was set by the change
sequence. The module type is used by windows driver not Linux driver.
It is defined in the following enum definition in cavs_ext_manifest.h.

UpDwMix module type is 5, Mux is 6 kpb is 11, and Selector is 12. The
module ASRC, GAIN, PROBE are type of efx (9) according to spec.

enum mod_type {
        ebasefw = 0,
        emixin,
        emixout,
        ecopier,
        epeakvol,
        eupdwmix,
        emux,
        esrc,
        ewov,
        efx,
        eaec,
        ekpb,
        emicselect,
        efxf,   /*i.e.SmartAmp */
        eaudclass,
        efakecopier,
        eiodriver,
        ewhm,
        egdbstub,
        esensing,
        emax,
        einvalid = emax
} ;

Signed-off-by: Rander Wang <rander.wang@intel.com>