Fix IMR type parsing
The IMR type can be specified either in the TOML configuration file
or on the command line. The command line value should override the
one from the configuration file. But the current code overwrites the
configuration file value with the default value even if no value has
been specified on the command line, which is wrong. Fix this by using
the default value when reading the configuration file and only
overwriting it when the respective command line parameter is used.

Signed-off-by: Guennadi Liakhovetski <guennadi.liakhovetski@linux.intel.com>