Fix regression - make default return code an error
A recent commit ec649f37d661 ("Convert all ELF addresses to cached
for calculations") had a bug: it assumed a call to
parse_uint32_hex_key() for an absent key would return an error by
default, which isn't the case. To force it set the default return
code to an error value.

Signed-off-by: Guennadi Liakhovetski <guennadi.liakhovetski@linux.intel.com>